refactor(settings): 拆分SettingsPage为主题+更新+Git+部署路径+Python环境
