+++ frontend/src/components/settings/index.ts
@@ -1,0 +1,6 @@
+export { CustomSelect } from "./CustomSelect"
+export { ThemeSettings } from "./ThemeSettings"
+export { UpdateCheckSection } from "./UpdateCheckSection"
+export { GitEnvironmentPanel } from "./GitEnvironmentPanel"
+export { DeploymentPathPanel } from "./DeploymentPathPanel"
+export { PythonEnvironmentPanel } from "./PythonEnvironmentPanel"
